+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - Add admin api endpoints for `BatchOrder` resources to create, read, list
   and cancel
 - Claim an order generated from a batch order through voucher code
+- Add description to order group model
 
 ### Changed